+++ src/app.d.ts
@@ -1,7 +1,7 @@
 // See https://kit.svelte.dev/docs/types#app
 
 import type { Session } from "@prisma/client"
-import type { EntireUser } from "$lib/server/database/auth"
+import type { EntireUser } from "$lib/server/database/auth/user"
 
 
 // for information about these interfaces
@@ -26,6 +26,7 @@ declare global {
 
         interface PageData {
             user: EntireUser | null
+            session: Session | null
         }
         // interface Platform {}
     }
