+++ src/routes/admin/+page.server.ts
@@ -8,9 +8,11 @@ export const load: PageServerLoad = async ({ cookies }) => {
     const loggedIn = Boolean(sessionToken && validate(sessionToken))
 
     const allPermissionGroups = await db.auth.permissionGroup.getAllPermissionGroups()
+    const allPermissions = await db.auth.permission.getAllPermissions()
 
     return {
         users: loggedIn ? await db.auth.user.getAllUsers() : [],
-        allPermissionGroups
+        allPermissionGroups,
+        allPermissions
     }
 }
